refactor(codegen-new): roteamento de classe registry data-driven (drena Date)
Remove o nome `Date` do CONTROLE DE FLUXO do dispatch (doutrina
PRIMORDIAL-vs-Registry). Os 4 literais `class == "Date"` (ctor, metodo de
instancia, static, instanceof) em globalclass.rs/call.rs viram o predicado
`is_pure_registry_class` = `classes.get(c).is_none() && class_meta(c).is_none()
&& registry::has_class(c)`. Hoje o conjunto = {Date}; Map/Set entram depois so
removendo a entrada class_meta/ambiente, sem branch novo.

dateclass.rs -> registryclass.rs: emit_registry_ctor / try_registry_instance_method
/ try_registry_static_call genericos sobre qualquer classe registry (resolvem via
o Registry REAL: class_ctor/class_member/class_static + emit_registry_call). A
unica logica Date-especifica restante fica isolada em registry_class_policy (o
unico literal "Date"), documentada como proximo drain -> metadata no spec Date de
rts-shared (flag divergent por-metodo + default-arg/overload do ctor).

Behavior-preserving. cargo test --release -p rts-codegen-new: 703/703 (baseline
703). Testes Date: 25/25. measure_new: 297/630 (>= 274 previo). Gate check-3
(Date no codegen): 24 -> 4.
