+++ crates/rts-codegen-new/src/front/run/mod.rs
@@ -28,7 +28,6 @@ mod call;
 mod call_shape;
 mod call_spread;
 mod class;
-mod dateclass;
 mod desugar;
 mod engineobj;
 mod expr;
@@ -50,6 +49,7 @@ mod optchain_lower;
 mod regex;
 pub(crate) mod registry;
 mod registry_call;
+mod registryclass;
 mod sig;
 mod stmt;
 mod thunk;
